+++ Kventin.Backend/Kventin.Domain/Domain/Message.cs
@@ -10,7 +10,7 @@ namespace Kventin.DataAccess.Domain
         /// <summary>
         /// Id отправителя
         /// </summary>
-        public int SenderId { get; set; }
+        public long SenderId { get; set; }
 
         /// <summary>
         /// Отправитель
