+++ Kventin.Backend/Kventin.Services/Interfaces/Services/IUserService.cs
@@ -4,8 +4,8 @@ namespace Kventin.Services.Interfaces.Services
 {
     public interface IUserService
     {
-        public Task<List<UserShortInfoDto>> GetUsersChildren(int parentId);
-        public Task SetChildrenForParent(int parentId, List<int> childrenIds);
+        public Task<List<UserShortInfoDto>> GetUsersChildren(long parentId);
+        public Task SetChildrenForParent(long parentId, List<long> childrenIds);
         public Task<List<UserShortInfoDto>> GetAllUsersShortInfoByRole(string rolename);
     }
 }
