+++ Kventin.Backend/Kventin.Services/Interfaces/Tools/IJwtProvider.cs
@@ -2,14 +2,14 @@
 {
     public interface IJwtProvider
     {
-        public string GenerateToken(int userId, string userLogin, List<string> rolenames, int selectedChildId = 0);
+        public string GenerateToken(long userId, string userLogin, List<string> rolenames, long selectedChildId = 0);
 
-        public int GetUserIdByToken(string token);
+        public long GetUserIdByToken(string token);
 
         public List<string> GetUserRolesByToken(string token);
 
         public string GetUserLoginByToken(string token);
 
-        public int GetChildIdByToken(string token);
+        public long GetChildIdByToken(string token);
     }
 }
